src/text-training/requirements.txt
torch>=2.2
transformers>=4.40
peft>=0.11
accelerate>=0.30
pyyaml>=6.0
